final transformation adjustments

+++ User/User.h
@@ -3,24 +3,22 @@
 
 #pragma once
 
-#include <iostream>
-
 class user
 {
 public:
 	float x_transformation;
 	float y_transformation;
 	float z_transformation;
-	
+
 	float x_rotation;
 	float y_rotation;
 	float z_rotation;
-	
+
 	float x_scale;
 	float y_scale;
 	float z_scale;
 
-	int rotation_angle;
+	float rotation_angle;
 
 	void userInput();
 };
